WIP - Refactor motoman_ma1400_support package, port meshes, new URDF This commit refactors the motoman_ma1400_support package to make it more consistent with the conventions of other motoman support packages. The URDF meshes and file definitions are now contained in this package, and a new URDF is implemented based on the URDF from the ma1400_sim package where it originally was. NOTE!! the origin for the inertial links of the new URDF are not set properly, need to figure out how to figure out proper values for this
