+++ motoman/package.xml
@@ -17,7 +17,7 @@
 
   <exec_depend>motoman_driver</exec_depend>
   <exec_depend>motoman_gp12_support</exec_depend>
-  <exec_depend>motoman_ma1400_support</exec_depend>
+  <exec_depend>motoman_ma1440_support</exec_depend>
   <exec_depend>motoman_gp7_support</exec_depend>
   <exec_depend>motoman_gp8_support</exec_depend>
   <exec_depend>motoman_mh12_support</exec_depend>
