--- motoman_ma1400_support/config/joint_names_ma1400.yaml
@@ -1,1 +1,0 @@
-controller_joint_names: ['joint_s', 'joint_l', 'joint_u', 'joint_r', 'joint_b', 'joint_t']
