--- motoman_ma1400_support/package.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format2.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="2">
- <name>motoman_ma1400_support</name>
- <version>0.3.5</version>
- <description>
-  <p>
-      ROS Industrial support for the Motoman MA1400 plus MotoMount Positioner.
-    </p>
-  <p>
-      This package contains configuration data, 3D models and launch files
-      for Motoman MA1400 robot plus MotoMount Positioner.
-    </p>
-  <p>
-   <b>Specifications</b>
-  </p>
-  <ul>
-   <li>MA1400 - Default</li>
-  </ul>
-  <p>
-      Joint limits and maximum joint velocities are based on the information 
-      found in the online 
-      https://cdn2.hubspot.net/hubfs/366775/Blog_Robots/MA1440.pdf
-      All urdfs are based on the default motion and joint velocity limits, 
-      unless noted otherwise.
-    </p>
-  <p>
-      Before using any of the configuration files and / or meshes included
-      in this package, be sure to check they are correct for the particular
-      robot model and configuration you intend to use them with.
-    </p>
- </description>
- <author email="gmattklein@gmail.com">Matthew Klein</author>
- <maintainer email="gmattklein@gmail.com">Matthew Klein</maintainer>
- <license>BSD</license>
- <buildtool_depend>catkin</buildtool_depend>
- <test_depend>roslaunch</test_depend>
- <exec_depend>motoman_driver</exec_depend>
- <exec_depend>robot_state_publisher</exec_depend>
- <exec_depend>rviz</exec_depend>
- <exec_depend>joint_state_publisher</exec_depend>
- <export>
-  <architecture_independent/>
- </export>
-</package>
