+++ motoman_ma1440_support/config/joint_names_ma1440.yaml
@@ -1,0 +1,1 @@
+controller_joint_names: ['joint_s', 'joint_l', 'joint_u', 'joint_r', 'joint_b', 'joint_t']
