+++ motoman_ma1440_support/package.xml
@@ -1,0 +1,45 @@
+<?xml version="1.0"?>
+<?xml-model href="http://download.ros.org/schema/package_format2.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
+<package format="2">
+ <name>motoman_ma1440_support</name>
+ <version>1.0.0</version>
+ <description>
+  <p>
+      ROS Industrial support for the Motoman MA1440 arm.
+    </p>
+  <p>
+      This package contains configuration data, 3D models and launch files
+      for Motoman MA1440 robot.
+    </p>
+  <p>
+   <b>Specifications</b>
+  </p>
+  <ul>
+   <li>MA1440 - Default</li>
+  </ul>
+  <p>
+      Joint limits and maximum joint velocities are based on the information 
+      found in the online 
+      https://cdn2.hubspot.net/hubfs/366775/Blog_Robots/MA1440.pdf
+      All urdfs are based on the default motion and joint velocity limits, 
+      unless noted otherwise.
+    </p>
+  <p>
+      Before using any of the configuration files and / or meshes included
+      in this package, be sure to check they are correct for the particular
+      robot model and configuration you intend to use them with.
+    </p>
+ </description>
+ <author email="ben@path-robotics.com">Ben Kent</author>
+ <maintainer email="ben@path-robotics.com">Ben Kent</maintainer>
+ <license>BSD</license>
+ <buildtool_depend>catkin</buildtool_depend>
+ <test_depend>roslaunch</test_depend>
+ <exec_depend>motoman_driver</exec_depend>
+ <exec_depend>robot_state_publisher</exec_depend>
+ <exec_depend>rviz</exec_depend>
+ <exec_depend>joint_state_publisher</exec_depend>
+ <export>
+  <architecture_independent/>
+ </export>
+</package>
