+++ motoman_ma1440_support/test/launch_test.xml
@@ -1,0 +1,43 @@
+
+<!--
+  launch_test.xml - ROSlaunch tests
+
+  Usage Instructions:
+
+  1. Add the following to your CMakeLists.txt:
+
+    find_package(roslaunch)
+    roslaunch_add_file_check(test/launch_test.xml)
+
+  2. Create a test directory under your package
+  3. Add the "launch_text.xml" file and fill out the test below.  Use the
+     following conventions:
+    a. Encapsulate each launch file test in it's own namespace.  By
+       convention the namespace should have the same name as the launch
+       file (minus ".launch" extension)
+    b. Create tests for each possible combination of parameters.  Utilize
+       sub-namespaces under the main namespace.
+
+  Notes:
+
+  1. XML extension is used in order to avoid being included
+  in roslaunch auto-complete.
+
+  2. Group tags with namespaces are used to avoid node name collisions when
+  testing multiple launch files
+-->
+<launch>
+    <arg name="req_arg" value="default"/>
+    <arg name="dx100" value="dx100"/>
+    <arg name="fs100" value="fs100" />
+
+<!--    <group ns="">-->
+        <include file="$(find motoman_ma1440_support)/launch/load_ma1440.launch"/>
+<!--    </group>-->
+
+    <rosparam command="load" file="$(find motoman_ma1440_support)/config/joint_names_ma1440.yaml" />
+    <rosparam command="load" file="$(find motoman_ma1440_support)/config/ma1440_motion_interface.yaml" />
+    <node name="joint_state_publisher" pkg="joint_state_publisher" type="joint_state_publisher" />
+    <node name="robot_state_publisher" pkg="robot_state_publisher" type="robot_state_publisher" />
+    <node name="rviz" pkg="rviz" type="rviz" args="-d $(find industrial_robot_client)/config/robot_state_visualize.rviz" required="true" />
+</launch>
